requirements.txt
python-docx==1.1.2
pandas==2.1.4
streamlit==1.31.0
